Moved localWarning from Header to state/application
