+++ src/custom/state/application/localWarning.ts
